refactor(biblio-ref): refactored code
